// Module5/case_study_angular/case-study/src/app/model/facility/Facility.ts
export interface Facility {
  id?: number;
  nameService?: string;
  area?: number;
  cost?: number;
  maxPeople?: number;
  standardRoom?: string;
  descriptionOtherConvenience?: string;
  areaPool?: number;
  numberOfFloors?: number;
  freeService?: string;
  rentType?: string;
  facilityType?: string;
}